Add emails to file
